Add Price Calculation Kit

+++ OrderManagementTool/OrderManagementTool.cs
@@ -13,6 +13,7 @@ namespace OrderManagementTool
         private OrderCreationPage frmOrderCreation;
         private UndoneOrdersPage frmUndoneOrders;
         private OrderDetailsPage frmOrderDetail;
+        private CalculatePriceKitPage frmPriceKit;
 
         // define delegate
         public delegate void DlgSendOperation(string operation, string orderNo);
@@ -144,20 +145,30 @@ namespace OrderManagementTool
 
         private void btnDeleteOrder_Click(object sender, EventArgs e)
         {
-            string orderNo = dgvTransaction.CurrentRow.Cells["OrderNo"].Value.ToString();
 
-            Order objOrder = new OrderManage().GetOrderByOrderNo(orderNo);
-            objOrder.User = new UserInfoManage().GetUserByOrderNo(orderNo);
-            new TransactionManage().DeleteTransactionRecord(orderNo);
-            new ItemManage().DeleteItemListByOrderNo(orderNo);
-            new OrderManage().DeleteOrder(objOrder);
-            int result = new UserInfoManage().DeleteUser(objOrder.User);
-            if (result > 0)
+            if (MessageBox.Show(this, "Delete?", "Prompt", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
             {
-                MessageBox.Show("Delete data sucessfully!");
-            }
+                string orderNo = dgvTransaction.CurrentRow.Cells["OrderNo"].Value.ToString();
 
-            ShowTransaction(tbSearch.Text.Trim(), Convert.ToInt32(cmbSorting.SelectedIndex));
+                Order objOrder = new OrderManage().GetOrderByOrderNo(orderNo);
+                objOrder.User = new UserInfoManage().GetUserByOrderNo(orderNo);
+                new TransactionManage().DeleteTransactionRecord(orderNo);
+                new ItemManage().DeleteItemListByOrderNo(orderNo);
+                new OrderManage().DeleteOrder(objOrder);
+                int result = new UserInfoManage().DeleteUser(objOrder.User);
+                if (result > 0)
+                {
+                    MessageBox.Show("Delete data sucessfully!");
+                }
+
+                ShowTransaction(tbSearch.Text.Trim(), Convert.ToInt32(cmbSorting.SelectedIndex));                
+            }
+        }
+
+        private void btnPriceKit_Click(object sender, EventArgs e)
+        {
+            frmPriceKit = new CalculatePriceKitPage();
+            frmPriceKit.Show();
         }
 
     }
